feat: combine personal/address/other into single patient info page
Merge the three Add Patient form steps (Personal, Address, Other) into
one combined page with section headers. Step indicator removed since
only one form step remains. Privacy Policy and Terms gates kept as-is;
flow is now Privacy -> Terms -> Patient Info -> Preview.
